[serve][llm] Move _maybe_setup_kv_aware_routing into kv_aware/utils.py
