+++ python/ray/llm/_internal/serve/core/server/builder.py
@@ -16,14 +16,9 @@ from ray.llm._internal.serve.core.configs.llm_config import (
 )
 from ray.llm._internal.serve.core.server.llm_server import LLMServer
 from ray.llm._internal.serve.observability.logging import get_logger
-from ray.llm._internal.serve.routing_policies.kv_aware.kv_aware_actor import (
-    KV_ROUTER_ACTOR_NAME,
-    KVRouterActor,
+from ray.llm._internal.serve.routing_policies.kv_aware.utils import (
+    _maybe_setup_kv_aware_routing,
 )
-from ray.llm._internal.serve.routing_policies.kv_aware.kv_aware_router import (
-    KVAwareRouter,
-)
-from ray.serve.config import DeploymentActorConfig, RequestRouterConfig
 from ray.serve.deployment import Application
 
 logger = get_logger(__name__)
@@ -41,32 +36,6 @@ DEFAULT_DEPLOYMENT_OPTIONS = {
 
 def _get_deployment_name(llm_config: LLMConfig) -> str:
     return llm_config.model_id.replace("/", "--").replace(".", "_")
-
-
-def _maybe_setup_kv_aware_routing(deployment_options: dict) -> None:
-    """Set up KV-aware routing when the deployment's request router is a
-    KVAwareRouter.
-
-    Currently attaches the KVRouterActor, which owns the deployment's global KV
-    radix tree for KV-aware request scoring.
-    """
-    request_router_config = deployment_options.get("request_router_config")
-    if isinstance(request_router_config, dict):
-        request_router_config = RequestRouterConfig(**request_router_config)
-    if not isinstance(request_router_config, RequestRouterConfig):
-        return
-    if not issubclass(request_router_config.get_request_router_class(), KVAwareRouter):
-        return
-
-    # TODO (jeffreywang): KVRouterActor requires init_kwargs such as block_size.
-    deployment_options["deployment_actors"] = [
-        *deployment_options.get("deployment_actors", []),
-        DeploymentActorConfig(
-            name=KV_ROUTER_ACTOR_NAME,
-            actor_class=KVRouterActor,
-            actor_options={"num_cpus": 0},
-        ),
-    ]
 
 
 def build_llm_deployment(
